refactor: remove the legacy genetic-algorithm engine
OR-Tools CP-SAT is now the only solver. The GA was a selectable alternative behind
a `WakfuSolver` toggle; it is removed entirely:

- delete genetic/{GeneticAlgorithm,Selection}.kt and genetic/wakfu/{Population,
  Cross,Mutation}.kt; keep GeneticAlgorithmResult (the shared streaming type, now
  produced only by the CP-SAT solver) in its own file.
- WakfuBestBuildFinderAlgorithm.run() goes straight to WakfuBuildSolver.optimize;
  drop the WakfuSolver enum and the `solver` / `populationSize` params + GA helpers.
- GUI: drop the solver field, the setSolver/restoredSolver plumbing and the solver
  toggle in the constraints panel.
- history: drop RequestSnapshot.solver — old saves keep parsing (ignoreUnknownKeys).
- tests: the GA was only a secondary oracle in WakfuBuildSolverTest; keep the
  stronger brute-force exhaustive-optimum assertions, drop the GA comparisons and
  helpers, and remove the GA-only GeneticAlgorithmZenithE2ETest.
- docs: AGENTS.md §4 / CLAUDE.md now describe a single engine.
