+++ CLAUDE.md
@@ -7,10 +7,10 @@ it is imported below so both files stay in sync.
 
 ## Claude Code working notes (project-specific)
 
-- **Engine.** The default solver is the Google OR-Tools CP-SAT solver
-  (`autobuilder/.../WakfuBuildSolver.kt`, deterministic & optimal); the original genetic algorithm
-  is still selectable via the `WakfuSolver` parameter. Both stream the same
-  `Flow<GeneticAlgorithmResult<BuildCombination>>`. See `AGENTS.md` §4.
+- **Engine.** The solver is the Google OR-Tools CP-SAT solver
+  (`autobuilder/.../WakfuBuildSolver.kt`, deterministic & optimal). It streams its result as a
+  `Flow<GeneticAlgorithmResult<BuildCombination>>` (legacy type name). The original genetic-algorithm
+  engine has been **removed** — OR-Tools is the only solver (no `WakfuSolver` toggle). See `AGENTS.md` §4.
 - **OR-Tools is native.** It loads a native library at runtime, so running/testing the engine needs
   extra JVM args (`--enable-native-access=ALL-UNNAMED`, `--add-opens …`) — already wired in the
   `autobuilder` and `gui-compose` build scripts. The first search pays a one-time cold start; the
